+++ app/helpers/events_helper.rb
@@ -1,2 +1,5 @@
 module EventsHelper
+  def event_params
+    params.require(:event).permit(:name, :description, :date, :user_id)
+  end
 end
