+++ app/views/users/show.html.erb
@@ -6,7 +6,7 @@
   <ul class="events">
     <%= label_tag 'Upcoming'%>
     <% @user.attended_events.sort{ |a,b| a.date <=> b.date }.each do |e| %>
-      <% if e.date > @date %>
+      <% if e.date >= @date %>
       <li>
         <%= link_to  e.name , event_path(e.id) %>
         <%= e.date %>
